Refactor: Divide form components for inner functions
