+++ src/app/_components/(auth)/clientComponents/forms/RegisterTermsForm.tsx
