+++ src/stories/(auth)/register/password/clientComponents.stories.tsx
@@ -1,8 +1,8 @@
 import { JSX, useState } from "react";
-import { RegisterFormProvider } from "@/app/_contexts/RegisterFormProvider"; // Correct import path
+import { RegisterFormProvider } from "@/contexts/RegisterFormProvider"; // Correct import path
 
-import { RegisterPasswordForm } from "@/app/_components/(auth)/clientComponents/RegisterPasswordForm";
-import { PasswordInput } from "@/app/_components/(auth)/clientComponents/ResetPasswordForm";
+import { RegisterPasswordForm } from "@/components/(auth)/clientComponents/forms/RegisterPasswordForm";
+import { ResetPasswordInputCC } from "@/components/(auth)/clientComponents/ResetPasswordInputCC";
 
 export default {
   title: "Components/Register/Password/clientComponents", // Updated title for uniqueness
@@ -23,7 +23,7 @@ export const DefaultRegisterPasswordForm = {
 
 export const DefaultPasswordInput = {
   render: () => (
-    <PasswordInput
+    <ResetPasswordInputCC
       password={""}
       setPassword={function (password: string): void {
         throw new Error("Function not implemented.");
